requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
openai==1.77.0
numpy==2.2.5
python-dotenv==1.1.0
pydantic==2.11.3
